move uniform.h to gl/ + cleanups
