+++ core/src/style/style.h
@@ -1,8 +1,8 @@
 #pragma once
 
 #include "gl.h"
+#include "gl/uniform.h"
 #include "data/tileData.h"
-#include "util/uniform.h"
 
 #include <memory>
 #include <string>
